Remove some out-of-scope functions from AuthUseCase and used the UserUseCase
