+++ app/src/main/java/com/weatherxm/usecases/WidgetCurrentWeatherUseCase.kt
@@ -7,6 +7,5 @@ import com.weatherxm.ui.common.UIDevice
 interface WidgetCurrentWeatherUseCase {
     suspend fun getUserDevice(deviceId: String): Either<Failure, UIDevice>
     fun removeWidgetId(widgetId: Int)
-
     fun getWidgetDevice(widgetId: Int): String?
 }
