+++ app/src/test/java/com/weatherxm/ui/login/LoginViewModelTest.kt
@@ -17,6 +17,7 @@ import com.weatherxm.data.models.User
 import com.weatherxm.data.network.AuthToken
 import com.weatherxm.ui.InstantExecutorListener
 import com.weatherxm.usecases.AuthUseCase
+import com.weatherxm.usecases.UserUseCase
 import com.weatherxm.util.Resources
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
@@ -37,6 +38,7 @@ import org.koin.dsl.module
 @OptIn(ExperimentalCoroutinesApi::class)
 class LoginViewModelTest : BehaviorSpec({
     val usecase = mockk<AuthUseCase>()
+    val userUseCase = mockk<UserUseCase>()
     val analytics = mockk<AnalyticsWrapper>()
     lateinit var viewModel: LoginViewModel
 
@@ -67,7 +69,7 @@ class LoginViewModelTest : BehaviorSpec({
         justRun { analytics.trackEventFailure(any()) }
         justRun { analytics.setUserId(user.id) }
         every { usecase.isLoggedIn() } returns true
-        every { usecase.shouldShowAnalyticsOptIn() } returns true
+        every { userUseCase.shouldShowAnalyticsOptIn() } returns true
         every {
             resources.getString(R.string.error_login_invalid_username)
         } returns invalidUsername
@@ -78,7 +80,7 @@ class LoginViewModelTest : BehaviorSpec({
             resources.getString(R.string.error_login_invalid_credentials)
         } returns invalidCredentials
 
-        viewModel = LoginViewModel(usecase, resources, analytics)
+        viewModel = LoginViewModel(usecase, userUseCase, resources, analytics)
     }
 
     context("Get if the user is logged in already or not") {
@@ -149,7 +151,7 @@ class LoginViewModelTest : BehaviorSpec({
             }
             When("login is a success and fetching the user is a failure") {
                 coMockEitherRight({ usecase.login(username, password) }, authToken)
-                coMockEitherLeft({ usecase.getUser() }, failure)
+                coMockEitherLeft({ userUseCase.getUser() }, failure)
                 runTest { viewModel.login(username, password) }
                 then("LiveData of login posts a success") {
                     viewModel.onLogin().isSuccess(Unit)
@@ -163,7 +165,7 @@ class LoginViewModelTest : BehaviorSpec({
             }
             When("login is a success and fetching the user is a success also") {
                 coMockEitherRight({ usecase.login(username, password) }, authToken)
-                coMockEitherRight({ usecase.getUser() }, user)
+                coMockEitherRight({ userUseCase.getUser() }, user)
                 runTest { viewModel.login(username, password) }
                 then("LiveData of login posts a success") {
                     viewModel.onLogin().isSuccess(Unit)
